pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
description = "A simple voice AI assistant built with LiveKit Agents"
requires-python = ">=3.9,<3.14"
dependencies = [
    "aiohttp>=3.13.2",
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
]
